Small fixes... also, unified UtilsTest with StringUtilTest.
