+++ src/javasources/KTool/test/org/kframework/utils/StringUtilTest.java
@@ -1,11 +1,45 @@
-// Copyright (c) 2012-2014 K Team. All Rights Reserved.
+// Copyright (c) 2014 K Team. All Rights Reserved.
 package org.kframework.utils;
-
 
 import junit.framework.Assert;
 import org.junit.Test;
 
 public class StringUtilTest {
+
+    @Test
+    public void StringUtilUnquote() throws Exception {
+        Assert.assertTrue(StringUtil.unquoteKString("\"\\n\"").equals("\n"));
+        Assert.assertTrue(StringUtil.unquoteKString("\"\\r\"").equals("\r"));
+        Assert.assertTrue(StringUtil.unquoteKString("\"\\f\"").equals("\f"));
+        Assert.assertTrue(StringUtil.unquoteKString("\"\\t\"").equals("\t"));
+
+        Assert.assertTrue(StringUtil.unquoteKString("\"\\x20\"").equals(" "));
+        Assert.assertTrue(StringUtil.unquoteKString("\"\\u0020\"").equals(" "));
+        Assert.assertTrue(StringUtil.unquoteKString("\"\\U00000020\"").equals(" "));
+
+        Assert.assertTrue(StringUtil.unquoteKString("\"\\U00000020\"").equals(" "));
+
+        try {
+            StringUtil.unquoteKString("\"\\U00110000\"");
+            throw new AssertionError();
+        } catch (IllegalArgumentException e) {
+        }
+        try {
+            StringUtil.unquoteKString("\"\\U0000d801\"");
+            throw new AssertionError();
+        } catch (IllegalArgumentException e) {
+        }
+    }
+
+    @Test
+    public void StringUtilEscapeShell() throws Exception {
+        String[] cmd1 = new String[] {"foo", "bar \" baz"};
+        String[] cmd2 = new String[] {"'"};
+        Assert.assertEquals("\"foo\" \"bar \\\" baz\"", StringUtil.escapeShell(cmd1, OS.WIN));
+        Assert.assertEquals("'foo' 'bar \" baz'", StringUtil.escapeShell(cmd1, OS.UNIX));
+        Assert.assertEquals("''\\'''", StringUtil.escapeShell(cmd2, OS.UNIX));
+    }
+
     @Test
     public void testBijection() {
         char[] all = new char[256];
@@ -21,15 +55,10 @@ public class StringUtilTest {
             Assert.assertEquals("Different values at position: " + i, all[i] , all2[i]);
     }
 
-    @Test
+    @Test(expected=IllegalArgumentException.class)
     public void testErrors() {
-        try {
-            String str = String.valueOf(new char[]{400});
-            StringUtil.enquoteCString(str);
-            throw new AssertionError("An " + IllegalArgumentException.class.getName() + " should have been thrown here.");
-        } catch (IllegalArgumentException e) {
-            // good
-        }
+        String str = String.valueOf(new char[]{400});
+        StringUtil.enquoteCString(str);
     }
 
     @Test
